refactor(db): Split lensed_auto_commit_fetcher.rs into directory module
